// src/app/xyz/types.d.ts
export declare type ModelType = 'CLM' | 'TFJS';
export declare type Vector2D = {
  y: number;
  x: number;
};
export declare type FacePosition = {
  nose: Vector2D,
  leftEye: Vector2D,
  rightEye: Vector2D,
  leftEar: Vector2D,
  rightEar: Vector2D,
};
export declare type FaceScale = {
  nose: number,
  leftEye: number,
  rightEye: number,
  leftEar: number,
  rightEar: number,
};